refactor(ui-build): rename build package to ui-build-sessions
