+++ plugins/shape-plugin/src/ui/components/build-progress/shapeBuildProgressSummaryComputationHelpers.ts
@@ -1,4 +1,4 @@
-export type { TaskCountSummary } from '../../../../../../packages/ui/build';
+export type { TaskCountSummary } from '../../../../../../packages/ui/build-sessions';
 export type { BuildStatus } from '@hierarchidb/components/build-status';
 export type { StageCountInfo } from './shapeBuildProgressSummaryCountHelpers.js';
 export {
