+++ plugins/shape-plugin/src/ui/components/build-progress/shapeBuildProgressSummaryCountHelpers.ts
@@ -1,9 +1,9 @@
-import type { TaskCountSummary } from '../../../../../../packages/ui/build';
+import type { TaskCountSummary } from '../../../../../../packages/ui/build-sessions';
 import type { BuildProgress } from './shapeBuildProgressMapping.js';
 import type { BuildStatus } from '@hierarchidb/components/build-status';
 import type { BuildStage } from '@hierarchidb/components/build-stage';
 import type { BuildTaskSummary } from '../../../../../../packages/build-api';
-import type { TaskStageCarrier } from '../../../../../../packages/ui/build';
+import type { TaskStageCarrier } from '../../../../../../packages/ui/build-sessions';
 
 export type StageCountInfo = { counts: TaskCountSummary; hasIncomplete: boolean };
