+++ plugins/shape-plugin/src/ui/components/build-progress/shapeBuildProgressSummaryPaneHelpers.ts
@@ -2,7 +2,7 @@ import type { BuildStatus } from '@hierarchidb/components/build-status';
 import type { BuildProgressStatus } from './shapeBuildProgressMapping.js';
 import type { BuildStage } from '@hierarchidb/components/build-stage';
 import type { BuildTaskSummary } from '../../../../../../packages/build-api';
-import type { TaskStageCarrier } from '../../../../../../packages/ui/build';
+import type { TaskStageCarrier } from '../../../../../../packages/ui/build-sessions';
 import { resolveMostAdvancedStageId } from './stagePriority.js';
 import type { StageCountInfo } from './shapeBuildProgressSummaryCountHelpers.js';
