+++ plugins/shape-plugin/src/ui/components/build-progress/useBuildProgress/useBuildProgress.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 import type { NodeType } from '@hierarchidb/core-types';
 import type { BuildUnifiedProgressInfo } from '../../../../../../../packages/build-api';
-import { usePluginBuildProgress } from '../../../../../../../packages/ui/build';
+import { usePluginBuildProgress } from '../../../../../../../packages/ui/build-sessions';
 import {
   toShapeProgress,
   toShapeStatus,
