+++ plugins/shape-plugin/src/ui/components/build-progress/useShapeBuildLabels/useShapeBuildLabels.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 import type { BuildStage } from '@hierarchidb/components/build-stage';
 import type { BuildStatus } from '@hierarchidb/components/build-status';
-import type { TaskCountSummary } from '../../../../../../../packages/ui/build';
+import type { TaskCountSummary } from '../../../../../../../packages/ui/build-sessions';
 import type { BuildProgress, BuildProgressStatus } from '~/ui/components/build-progress/shapeBuildProgressMapping';
 import { formatTaskDisplayMessage } from '~/ui/components/build-progress/taskDisplayText';
